refactor: split PurchaseOrderForm into smaller components
- Extract types and interfaces to PurchaseOrderTypes.tsx
- Extract validation logic to PurchaseOrderValidation.tsx
- Extract helper functions to PurchaseOrderHelpers.tsx
- Create PurchaseOrderLineItem component for individual line items
- Create PurchaseOrderLines component for line items management
- Create PurchaseOrderHeader component for form header section
- Create MaterialResolver component for material resolution logic
- Refactor main PurchaseOrderForm.tsx to orchestrate sub-components
- Reduce main component from 801 lines to manageable size
- Improve maintainability, testability, and code organization
- All existing functionality preserved with no behavioral changes

Fixes #99
